requirements.txt

flask
flask-cors
pytesseract
Pillow
openai
